Simplify the query to retrieve ratings
The SQL is a lot more complex than it needs to be. Considering that the
ratings data aggregated with outer joins is not used anywhere in the
plugin's code, we might as well remove it.

This fixes compatibility issues with PostgreSQL, caused by earlier use
of the MySQL-specific IFNULL() function.

Fixes #4